% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusionMatrix.R
\name{confusionMatrix}
\alias{confusionMatrix}
\title{Generic function to produce a confusion matrix (related to a classification problem)}
\usage{
confusionMatrix(x, ...)
}
\arguments{
\item{x}{object (usually a model fit object) that contains all information 
needed to produce the confusion matrix.}

\item{...}{further arguments for a specific method}
}
\value{
A confusion matrix
}
\description{
Generic function to produce a confusion matrix (related to a classification problem)
}
\author{
Tobias Verbeke
}
\keyword{models}
