% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topTable.R
\docType{methods}
\name{topTable-methods}
\alias{topTable-methods}
\alias{topTable,glmnet-method}
\alias{topTable,lognet-method}
\alias{topTable,elnet-method}
\title{Methods for topTable}
\arguments{
\item{fit}{object resulting from a classification or regression procedure}

\item{n}{number of features that one wants to extract from a table that
ranks all features according to their importance in the classification
or regression model; defaults to 10 for limma objects}
}
\description{
Methods for topTable. topTable extracts the top n most important features
for a given classification or regression procedure
}
\section{Methods}{

\describe{glmnet and lognet
\itemize{
\item{fit = "glmnet", n = "numeric"}{glmnet objects are produced by \code{lassoClass} (a4Classif) or 
\code{lassoReg} (a4Base)}
\item{fit = "lognet", n = "numeric"}{lognet objects are produced by \code{lassoClass} (a4Classif) or 
  \code{lassoReg} (a4Base)}
\item{fit = "elnet", n = "numeric"}{elnet objects are produced by \code{lassoClass} (a4Classif) or 
  \code{lassoReg} (a4Base)}
}
}
}

\keyword{manip}
\keyword{methods}
