\name{confusionMatrix}
\alias{confusionMatrix}
\title{
  Generic function to produce a confusion matrix (related to a classification problem)
}
\description{
  Generic function to produce a confusion matrix (related to a classification problem)
}
\usage{
confusionMatrix(x, ...)
}
\arguments{
  \item{x}{object (usually a model fit object) that contains all information 
    needed to produce the confusion matrix. 
}
  \item{\dots}{
    further arguments for a specific method
}
}
\author{
  Tobias Verbeke
}
\keyword{models}
